feat: improve api usability
